feat: add Rotate PDF action and configurable JPEG/WebP export quality
PdfPipeline::rotate() additively rotates every page by a multiple of 90°
via the /Rotate entry, normalized into [0,360); three tests cover the
additive, accumulating, and validation paths. A new rotate Tauri command,
RotatePanel (90/180/270 picker), and matching wiring across api.js,
App.svelte, ActionSidebar, and ParamsPanel surface it in the GUI.

encode::save() and PdfPipeline::save_page_image() now take a quality: u8
argument so JPEG/WebP exports can be tuned (PNG/TIFF ignore it). The
ExportPanel shows a quality slider for those formats, exportQuality (90
default) is plumbed through runActionOnPaths and the export_images
command, and rbara + tests + doc examples are updated to thread the new
argument.

+++ rustybara/src/stream/filter.rs
@@ -906,7 +906,7 @@ mod tests {
     #[test]
     fn object_to_f64_negative() {
         assert!((object_to_f64(&Object::Integer(-7)) - (-7.0)).abs() < 1e-10);
-        assert!((object_to_f64(&Object::Real(-2.5)) - (-2.5 as f64)).abs() < 0.01);
+        assert!((object_to_f64(&Object::Real(-2.5)) - (-2.5_f64)).abs() < 0.01);
     }
 
     // ── operands_to_rect ────────────────────────────────────────────
